requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
black==23.12.0
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dj-database-url==2.1.0
Django==4.2.16
django-anymail==10.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==37.12.0
flake8==6.1.0
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==5.13.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.31.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==5.5.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0

# Reports dependencies
weasyprint==60.1
openpyxl==3.1.2
